requirements.txt
Django==5.2
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-ckeditor==6.7.0 